Define _GNU_SOURCE for a declaration of vasprintf
This is necessary to build with compilers which do not support
implicit function declarations (a language feature that was removed
from C99).

Defining _GNU_SOURCE during the build exposes that the imported
util-linux sources are not built with some HAVE_* macros defined,
for things that the glibc headers provide with _GNU_SOURCE.  These
HAVE_* macros need to be defined as well.