fix: fix fail to kill linglong app
Previously, when killing an application, system-monitor needed to
obtain the pid of the app. However, on a desktop environment using X11,
system-monitor obtained an inaccurate pid of the linglong app, which
resulted in failure to kill the app or accidental killing. Now consider
using the window manager to kill the application, because the window
manager contains detailed information about the application window.

Log: fix fail to kill linglong app
Bug: https://pms.uniontech.com/bug-view-283189.html